requirements.txt
asgiref==3.5.2
cfgv==3.3.1
distlib==0.3.6
Django==3.2.16
django-extensions==3.2.1
django-rest-authtoken==2.1.4
django-timestampable-model==0.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
filelock==3.8.0
identify==2.5.8
model-bakery==1.9.0
nodeenv==1.7.0
platformdirs==2.5.3
pre-commit==2.20.0
PyJWT==2.5.0
pytz==2022.4
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
virtualenv==20.16.6
